Lots of work on the student-environment Docker image
Change Dockerfile and build script to take advantage of Docker
cache. Docker caches intermediate steps, so moving the "apt-get update
&& intall && clean" into its own RUN command greatly speeds up
rebuilds of our image.

Updated the README to use the coderdojotc.org Docker namespace.

Change the student environment to enable RCON for the Minecraft
server. Currently uses the same password as IPython, but this really
needs to be passed in from the environment so it doesn't get baked
into the image.

Also note, RCON does not currently work on CanaryMod. It looks like
there is a bug in the way CanaryMod handles the server
configuration. Need to work on fixing that.