Overhaul student-env-image build to use Supervisor and run unattended
Uses run scripts for Canary and IPython, launched by Supervisor.

Uses tmux to manage Canary/Minecraft.

Treats Canary as a group so the shutdown signals from Supervisor reach
the right processes.

Moved more logic into Docker file to speed builds and cut down on
extra scripts. Leverages information passed in environment variables
to better work with Docker.

During the image build, we pull in the CanaryMod and RaspberryJuice
jars. Change this in the Dockerfile if you need to update the
versions.

Starts off with the basic Canary config files in the image directory,
instead of launching Canary twice, once to create the files, and
another to run the game.